Restructing of the project
