+++ src/index.js
@@ -1,18 +1,20 @@
 import './scss/style.scss';
 import 'bootstrap';
 import { Modal } from 'bootstrap';
-import { Project, Todo } from './js/models';
-import { appendTodo, appendProject, getActiveTab } from './js/utils';
+import Project from './js/models/project';
+import Todo from './js/models/todo';
+import { appendTodo, appendProject, getActiveTab } from './js/dom';
+import { makeTodoForm } from './js/components';
 
 const projectForm = document.getElementById('projectForm');
-const projectsTree = document.getElementById('projectsTree');
+const todoModal = document.getElementById('todoModal');
 
 // Handle project form submission
 projectForm.onsubmit = (event) => {
   event.preventDefault();
 
-  const formData = new FormData(event.currentTarget);
-  const newProject = new Project(formData.get('projectName'));
+  const projectName = document.getElementById('projectName').value;
+  const newProject = new Project(projectName);
   const allProjects = Project.getAll();
 
   allProjects.push(newProject);
@@ -22,23 +24,38 @@ projectForm.onsubmit = (event) => {
 };
 
 // Handle todo form submission
-todoForm.onsubmit = (event) => {
-  event.preventDefault();
+todoModal.addEventListener('show.bs.modal', () => {
+  // Append form
+  todoModal.querySelector('.modal-body').appendChild(makeTodoForm());
 
-  const todoData = {
-    title: document.getElementById('title').value,
-    description: document.getElementById('description').value,
-    date: document.getElementById('date').value,
-    priority: document.getElementById('priority').value,
-    project: Project.get(getActiveTab().id),
+  const todoForm = document.getElementById('todoForm');
+
+  // Handle form submission
+  todoForm.onsubmit = (event) => {
+    event.preventDefault();
+
+    const todoData = {
+      title: document.getElementById('title').value,
+      description: document.getElementById('description').value,
+      date: document.getElementById('date').value,
+      priority: document.getElementById('priority').value,
+      project: Project.get(getActiveTab().id),
+    };
+
+    const newTodo = new Todo(todoData);
+
+    newTodo.save();
+    appendTodo(newTodo);
+
+    todoForm.reset();
+    Modal.getInstance(document.getElementById('todoModal')).hide();
   };
-  const newTodo = new Todo(todoData);
+});
 
-  newTodo.save();
-  appendTodo(newTodo);
-
-  Modal.getInstance(document.getElementById('taskModal')).hide();
-};
+// Remove form from modal body when modal is hidden
+todoModal.addEventListener('hide.bs.modal', () => {
+  todoModal.querySelector('.modal-body').innerHTML = '';
+});
 
 // Fetch all projects and display them on DOM
 const allProjects = Project.getAll();
